requirements.txt
streamlit
langchain
langchain-google-genai
langchain-experimental
langchain-community
pymysql
python-dotenv
sentence-transformers
chromadb
streamlit-option-menu
Pillow
altair==4.2.0